refactor: move extracted subcomponent to new file
