+++ src/views/components/MetadataFilters/MetadataFilters.tsx
@@ -1,0 +1,28 @@
+import { Genre, Platform, Publisher, Tag } from '@data/types'
+import React from 'react'
+import { FilterOptions } from '@views/components'
+
+
+interface IMetadataFilters {
+	metadataList: Map<string, Tag[] | Genre[] | Platform[] | Publisher[]>
+}
+
+export function MetadataFilters({ metadataList }: IMetadataFilters) {
+	const metadataFilters: React.ReactNode[] = []
+
+	metadataList.forEach((metadata, key) => {
+		metadataFilters.push(
+			<li key={key}>
+				<details>
+					<summary>{key}:</summary>
+
+					<FilterOptions options={metadata}
+								   typeName={key}
+					/>
+				</details>
+			</li>
+		)
+	})
+
+	return metadataFilters
+}
